+++ includes/plugin/class-core.php
@@ -12,13 +12,15 @@
 
 namespace KeysMaster\Plugin;
 
+use KeysMaster\Plugin\Feature\Analytics;
 use KeysMaster\System\Loader;
 use KeysMaster\System\I18n;
 use KeysMaster\System\Assets;
 use KeysMaster\Library\Libraries;
-
+use KeysMaster\System\Option;
 use KeysMaster\System\Nag;
 use KeysMaster\System\Password;
+use KeysMaster\System\Cache;
 
 /**
  * The core plugin class.
@@ -55,6 +57,9 @@ class Core {
 		$this->define_global_hooks();
 		$this->define_admin_hooks();
 		$this->define_public_hooks();
+		if ( \DecaLog\Engine::isDecalogActivated() && Option::network_get( 'metrics' ) ) {
+			$this->define_metrics();
+		}
 	}
 
 
@@ -111,6 +116,34 @@ class Core {
 		$plugin_public = new Keys_Master_Public();
 		$this->loader->add_action( 'wp_enqueue_scripts', $plugin_public, 'enqueue_styles' );
 		$this->loader->add_action( 'wp_enqueue_scripts', $plugin_public, 'enqueue_scripts' );
+	}
+
+	/**
+	 * Register all metrics of the plugin.
+	 *
+	 * @since  1.2.0
+	 * @access private
+	 */
+	private function define_metrics() {
+		$span      = \DecaLog\Engine::tracesLogger( POKM_SLUG )->start_span( 'Metrics collation' );
+		$cache_id  = 'metrics/lastcheck';
+		$analytics = Cache::get_global( $cache_id );
+		if ( ! isset( $analytics ) ) {
+			$analytics = Analytics::get_status_kpi_collection( [ 'site_id' => 0 ] );
+			Cache::set_global( $cache_id, $analytics, 'statistics' );
+		}
+		if ( isset( $analytics ) ) {
+			$metrics = \DecaLog\Engine::metricsLogger( POKM_SLUG );
+			if ( array_key_exists( 'data', $analytics ) ) {
+				foreach ( $analytics['data'] as $kpi ) {
+					$m = $kpi['metrics'] ?? null;
+					if ( isset( $m ) ) {
+						$metrics->createProdGauge( $m['name'], $m['value'], $m['desc'] );
+					}
+				}
+			}
+		}
+		\DecaLog\Engine::tracesLogger( POKM_SLUG )->end_span( $span );
 	}
 
 	/**